refactor: move tests to their respective packages
