+++ src/scripts/main.go
